Disabling walk in refrigeration  
Refrigeration is new to create_tyipcal from earlier versions so we didn't hit this before. Unlike SWH which doesn't have to be tied to a space and can survey geometry being replaced, that isn't the case for walk in refrigeration which requires a zone. My recommendation for now is not to include refrigeration when using urban_geometry.

I made an addition of __SKIP__ argument into blend, urban_geometry, and second create_typical instance to both test that refrigeration works when using _create_bar and also to provide users a path to easily switch between the two workflows with the same template OSW. I have left the __SKIP__ value to false for now, but if you turn them to true, and then set add refrigeration and hvac to true on the first instance of create_typical everything runs fine.

One last change was setting timestep to 4 per hour instead of 1. I hit warnings in EnergyPlus when it was set to 1 time step an hour.